+++ src/App.jsx
@@ -1,19 +1,12 @@
 import PropTypes from "prop-types";
 import { useEffect, useState } from "react";
-import {
-  RouterProvider,
-  createBrowserRouter,
-  useNavigation,
-} from "react-router-dom";
-import routes from "./routes/routes";
+import { RouterProvider } from "react-router-dom";
 import Loading from "./components/Loading";
-import { CartProvider } from "./components/Cart";
-import ScrollButtons from "./components/ScrollButtons";
+import CartProvider from "./context/CartContext";
 import styles from "./styles/App.module.css";
 
 // const App = () => {
 //   const router = createBrowserRouter(routes);
-//   console.log(router);
 //   return (
 //     <div id={styles.app}>
 //       <CartProvider>