+++ src/components/ErrorPage.jsx
@@ -3,10 +3,10 @@ import styles from "../styles/ErrorPage.module.css";
 
 const ErrorPage = () => {
   const { error, status } = useRouteError();
-  console.log(error);
-  console.log(status);
+
   return (
     <section id={styles["error-section"]} role="region">
+      {status && <h1>{status}</h1>}
       <p>{error.message}</p>
     </section>
   );