+++ src/components/Shop.jsx
@@ -1,15 +1,13 @@
-import { Outlet, useOutletContext, useRouteLoaderData } from "react-router-dom";
+import { Outlet, useRouteLoaderData } from "react-router-dom";
 import Categories from "./Categories";
 import styles from "../styles/Shop.module.css";
 
 const Shop = () => {
   const { categories, products } = useRouteLoaderData("root");
-  // const ScrollButtons = useOutletContext();
   return (
     <section id={styles["shop"]}>
       <Categories categories={categories} />
       <Outlet context={[products]} />
-      {/* <ScrollButtons /> */}
     </section>
   );
 };